TG-82: sidebar working, style done
