+++ src/Components/Shared/Sidebar/sidebar.module.css
@@ -1,0 +1,65 @@
+.Sidebar {
+    position: fixed;
+    display: flex;
+    flex-direction: column;
+    align-content:flex-end;
+    z-index: 200;
+    box-sizing: border-box;
+    padding: 5px 20px;
+    top: 0;
+    left: 0;
+    height: 100%;
+    width: 20vw;
+    background-color: rgba(175, 175, 194, 1);
+    color: rgba(55, 56, 103, 1);
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+}
+
+.sidebarTitle {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    font-size: 18px;
+    font-weight: bold;
+}
+
+.container{
+    display: flex;
+    flex-direction: column;
+    padding: 20px 0;
+    width: 100%;
+    box-shadow: 0px 1px rgba(55, 56, 103, 1)
+}
+
+.Sidebar a {
+    font-size: 15px;
+    width: 100%;
+    padding: 5px 0 5px 20px;
+    font-weight: lighter;
+    border-radius: 2px;
+}
+
+.Sidebar a:hover {
+    cursor:pointer;
+    background-color: rgb(60, 60, 80);
+    color:white;
+    
+}
+
+.Sidebar ul {
+    font-size: 12px;
+    padding: 5px 0 5px 20px;
+    font-weight: lighter;
+    border-radius: 2px;
+}
+
+.Sidebar ul:hover {
+    cursor:pointer;
+    background-color: rgb(60, 60, 80);
+    color:white;   
+}
+
+.sidebarTitle > i:hover {
+    cursor:pointer;
+    color:rgb(198, 46, 46);
+}
