--- src/Components/Shared/navbar.module.css
@@ -1,22 +1,0 @@
-.Sidebar {
-    display: flex;
-    flex-direction: column;
-    align-content: flex-start;
-    height: 100%;
-    width: 268px;
-    background-color: rgba(175, 175, 194, 1);
-    color: rgba(55, 56, 103, 1);
-}
-
-.Sidebar h2 {
-    font-size: 14px;
-    margin-left: 30px;
-    font-weight: bold;
-    margin-bottom: 20px;
-}
-
-.Sidebar a {
-    font-size: 12px;
-    margin-left: 48px;
-    font-weight: lighter;  
-}
