+++ src/Components/Tasks/tasks.module.css
@@ -3,5 +3,4 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  padding: 1%;
 }
